requirements.txt
opencv-contrib-python==4.8.0.74
numpy==1.24.3
streamlit==1.25.0
pillow>=7.1.0,<10.0.0
python-dotenv==1.0.0 